refactor(community): replace group action dialogs with confirm sheet and snackbars
