+++ assets/translations/tr.json
@@ -433,15 +433,12 @@
         "createdAtLabel": "Oluşturulma",
         "adminsTitle": "YÖNETİCİLER",
         "leaveGroup": "Gruptan ayrıl",
-        "leaveConfirmTitle": "Gruptan ayrıl",
-        "leaveConfirmMessage": "Bu gruptan ayrılmak istediğine emin misin?",
-        "leaveConfirmCancel": "Vazgeç",
-        "leaveConfirmApprove": "Ayrıl",
+        "leaveSuccess": "Gruptan ayrıldın.",
         "closeGroup": "Grubu kapat",
         "closeGroupConfirmTitle": "Grubu kapat",
-        "closeGroupConfirmMessage": "Bu grubu kapatmak istediğine emin misin? Bu işlem tüm üyeler için grubu kaldırır ve geri alınamaz.",
+        "closeGroupConfirmMessage": "Bu işlem tüm üyeler için grubu kaldırır ve geri alınamaz.",
         "closeGroupConfirmCancel": "Vazgeç",
-        "closeGroupConfirmApprove": "Kapat"
+        "closeGroupSuccess": "Grup kapatıldı."
       }
     },
     "createGroup": {
